adding light theme in header
