+++ packages/design/tailwind/css/components/header.css
@@ -1,8 +1,15 @@
 .gi-header {
-  @apply gi-grid gi-relative gi-overflow-x-clip gi-bg-color-surface-system-primary-default gi-w-full gi-font-primary;
+  @apply gi-grid gi-relative 
+  gi-overflow-x-clip
+  gi-w-full gi-font-primary;
 }
+
 .gi-header-secondary-bar {
-  @apply gi-hidden lg:gi-flex gi-bg-color-surface-system-primary-subtle gi-h-10 gi-items-center;
+  @apply gi-hidden 
+  gi-h-10
+  gi-items-center
+  
+  lg:gi-flex;
 }
 
 .gi-header-secondary-bar ul {
@@ -10,15 +17,44 @@
 }
 .gi-header-secondary-item,
 .gi-header-secondary-item-slot a {
-  @apply gi-border gi-border-solid gi-border-transparent focus-visible:gi-outline-offset-0 focus-visible:gi-outline-none focus-visible:gi-border focus-visible:gi-border-solid focus-visible:gi-border-color-shadow-intent-focus-default focus:gi-border focus:gi-border-solid focus:gi-border-color-shadow-intent-focus-default gi-block gi-text-white gi-stroke-white hover:gi-bg-black hover:gi-bg-opacity-20 gi-py-1 gi-px-2 gi-rounded-sm;
+  @apply gi-block
+  gi-px-2
+  gi-py-1
+  gi-border
+  gi-border-solid 
+  gi-border-transparent 
+  gi-rounded-sm
+  
+  focus-visible:gi-outline-offset-0 
+  focus-visible:gi-outline-none
+  focus-visible:gi-border
+  focus-visible:gi-border-solid
+  focus-visible:gi-border-color-shadow-intent-focus-default
+  
+  focus:gi-border
+  focus:gi-border-solid
+  focus:gi-border-color-shadow-intent-focus-default 
+  
+  
+  hover:gi-bg-black
+  hover:gi-bg-opacity-20;
 }
 
 .gi-header-secondary-item-slot > :first-child {
-  @apply gi-flex gi-items-center !gi-text-sm !gi-text-white;
+  @apply gi-flex
+  gi-items-center
+  !gi-text-sm;
 }
 
 .gi-header-menu {
-  @apply gi-h-20 gi-items-center gi-justify-between gi-flex gi-flex-row gi-flex-nowrap gi-bg-color-surface-system-primary-default gi-py-3 sm:gi-py-4;
+  @apply gi-h-20 gi-items-center 
+  gi-justify-between
+  gi-flex
+  gi-flex-row
+  gi-flex-nowrap
+  gi-py-3
+  
+  sm:gi-py-4;
 }
 
 .gi-header-logo {
@@ -30,13 +66,38 @@
 }
 
 .gi-header-title {
-  @apply gi-min-w-0 gi-heading-sm gi-tracking-wider gi-text-white gi-stroke-white gi-ml-4 md:gi-ml-6 lg:gi-ml-12 gi-grow lg:gi-block gi-truncate gi-overflow-hidden gi-text-ellipsis;
+  @apply gi-min-w-0 gi-heading-sm gi-tracking-wider
+  gi-truncate
+  gi-overflow-hidden
+  gi-text-ellipsis
+  gi-ml-4 
+  
+  md:gi-ml-6
+  lg:gi-ml-12
+  lg:gi-block
+  gi-grow;
 }
 .gi-header-tool-item {
-  @apply gi-border gi-border-solid gi-border-transparent gi-text-white gi-stroke-white gi-flex gi-rounded-sm hover:gi-bg-black hover:gi-bg-opacity-20 gi-p-2 gi-items-center gi-gap-md gi-cursor-pointer gi-whitespace-nowrap gi-overflow-hidden gi-text-ellipsis;
+  @apply gi-border 
+  gi-border-solid
+  gi-border-transparent
+  gi-flex gi-rounded-sm
+  
+  hover:gi-bg-black
+  hover:gi-bg-opacity-20
+  
+  gi-p-2
+  gi-items-center
+  gi-gap-md
+  gi-cursor-pointer
+  gi-whitespace-nowrap
+  gi-overflow-hidden
+  gi-text-ellipsis;
 }
 .gi-header-tool-item:first-child {
-  @apply lg:gi-flex gi-text-md gi-font-bold gi-text-white gi-stroke-white;
+  @apply lg:gi-flex
+  gi-text-md
+  gi-font-bold;
 }
 .gi-header-nav {
   @apply gi-hidden lg:gi-flex gi-gap-4 gi-text-white gi-stroke-white;
@@ -61,12 +122,11 @@
    focus:gi-border-solid focus-within:gi-border-solid;
 }
 
-.gi-header-separator {
-  @apply gi-hidden lg:gi-block gi-border-l-xs gi-border-solid gi-border-white gi-h-8 gi-mx-6;
-}
-
 .gi-header-divider {
-  @apply gi-border-l-xs gi-border-solid gi-border-white gi-h-8 gi-mx-2;
+  @apply gi-border-l-xs 
+  gi-border-solid    
+   gi-h-8 
+   gi-mx-2;
 }
 .gi-header-overlay {
   @apply gi-top-0 gi-z-900 gi-pointer-events-none gi-hidden gi-w-full gi-h-full gi-bg-black gi-opacity-20;
@@ -150,4 +210,12 @@
 a.gi-list-item:hover,
 .gi-list-item > a:hover {
   @apply hover:gi-underline hover:gi-underline-offset-sm;
+}
+
+.gi-header-slot-container {
+  @apply gi-border-b-2xl
+  gi-bg-gray-50
+  gi-py-4
+  gi-px-4  
+  gi-order-3;
 }
