+++ packages/react/ds/src/assets/logos/gov-of-ireland/harp-black.tsx
@@ -1,0 +1,193 @@
+function HarpBlack() {
+  return (
+    <svg
+      width="181"
+      height="64"
+      viewBox="0 0 181 64"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_209_1004)">
+        <path
+          d="M9.14178 17.6134C9.65474 17.6134 10.0707 17.1974 10.0707 16.6845C10.0707 16.1715 9.65474 15.7555 9.14178 15.7555C8.62882 15.7555 8.21282 16.1715 8.21282 16.6845C8.2133 17.1972 8.62882 17.613 9.14178 17.6134Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M7.66748 46.4128C7.0941 46.4126 6.62915 46.8776 6.62891 47.4509C6.62867 48.0243 7.09362 48.4893 7.667 48.4895C8.24038 48.4897 8.70533 48.0248 8.70557 47.4514C8.70486 46.8785 8.24062 46.4138 7.66748 46.4128Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M1.71489 43.8473C1.7039 43.9837 1.7156 44.1207 1.74975 44.2533C1.78056 44.3873 1.82355 44.5181 1.87799 44.6445C2.73913 44.2982 3.64302 43.8327 3.47418 41.9163C3.41089 41.4693 3.32492 41.0258 3.21627 40.5876C2.44898 37.4511 2.13137 34.2212 2.27274 30.9952C2.51155 26.2186 4.86953 21.8186 6.00315 19.7039C6.23909 19.2636 6.42583 18.9161 6.51634 18.706C6.85282 17.9298 5.44338 16.7262 4.86905 16.3239C4.60421 16.4879 4.41771 16.7532 4.35299 17.0579C4.30833 17.2531 4.31788 17.4565 4.38093 17.6464C2.68015 20.1937 1.44957 23.025 0.74772 26.0065C0.74772 26.0065 0.715242 26.1271 0.685152 26.2862C0.231896 28.305 0.00431353 30.3681 0.0067016 32.4374C-0.012403 32.4796 0.0143435 33.3732 0.024851 33.6173C-0.0424926 37.102 0.530405 40.5694 1.71489 43.8473Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M28.3136 20.6843C27.7911 20.6841 27.3672 21.1077 27.3672 21.6302C27.3669 22.1528 27.7906 22.5766 28.3131 22.5766C28.8356 22.5766 29.2595 22.1532 29.2595 21.6307C29.2588 21.1084 28.8359 20.685 28.3136 20.6843Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M22.5456 36.3213C23.0517 36.3213 23.4621 35.9109 23.4621 35.4047C23.4621 34.8985 23.0517 34.4882 22.5456 34.4882C22.0394 34.4882 21.629 34.8985 21.629 35.4047C21.629 35.9109 22.0394 36.3213 22.5456 36.3213Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M30.5133 18.4247C30.9699 16.7817 30.4894 15.6715 30.0051 15.0264C29.7191 14.6412 29.3539 14.3217 28.9341 14.0896C27.1027 12.8435 24.8651 13.6798 24.7982 13.7106C23.7154 14.4416 22.3712 14.6732 21.1062 14.3468C17.6946 13.4761 15.0095 9.1026 14.5054 8.23477L14.2229 7.89399C14.1412 7.80277 14.0316 7.74163 13.911 7.71966L13.2526 7.28981C13.0893 7.17853 12.8731 7.18402 12.7158 7.30342L12.6489 7.35309C12.6429 7.35309 12.6374 7.35023 12.6305 7.34927C12.2061 7.30987 11.782 7.4288 11.4401 7.6836C11.0102 7.96683 10.5923 8.4165 10.6393 8.85686L9.77604 9.50355C9.42667 9.53221 9.09186 9.65496 8.80673 9.8589C8.46404 10.0817 8.10248 10.4378 8.02201 10.8158L7.46129 11.2368L7.36935 11.3173L7.35884 11.367C7.32923 11.4135 7.30869 11.4651 7.29866 11.5193C7.28481 11.5523 7.27884 11.5881 7.28123 11.6237L7.28815 11.7634L7.32827 11.8131C7.39251 11.9413 7.81448 12.7845 8.09222 13.3089C7.2029 14.1185 6.38045 14.9985 5.63298 15.9406C6.35084 16.5075 7.74069 17.8045 7.21842 19.0097C7.11478 19.2485 6.93185 19.5912 6.67752 20.0653C5.57041 22.1305 3.26617 26.4283 3.03667 31.0334C2.89984 34.1909 3.21172 37.352 3.96324 40.4221C4.0805 40.8923 4.17172 41.3688 4.2362 41.8492C4.44205 44.1819 3.22223 44.9084 2.30664 45.2948C3.00252 46.8682 3.89112 48.3493 4.95143 49.7043C5.74761 50.6135 6.81365 52.3083 6.59657 53.3177C6.3463 53.2265 6.07645 53.2019 5.81376 53.2461C5.31012 53.3597 4.8891 53.7029 4.67632 54.1734C4.36946 54.8432 4.49889 55.632 5.00373 56.1686C5.45029 56.6023 6.11633 56.7016 6.92588 56.4337C6.93281 56.4501 6.94093 56.4659 6.94976 56.4814C7.03693 56.6398 7.19502 56.7465 7.3746 56.768C7.50595 56.7826 7.63777 56.79 7.77007 56.79C9.22655 56.79 11.5394 55.9231 14.331 51.8225C15.3051 51.9985 16.1503 52.1296 16.8328 52.2111L17.9361 52.4045C18.3984 52.5557 19.4033 52.4981 19.9189 51.1899L31.0475 22.4044C31.6653 21.0881 31.4573 19.5323 30.5155 18.4244L30.5133 18.4247ZM14.2895 10.193C14.3134 10.2345 16.7287 14.3786 21.3565 15.3261C21.6605 15.3897 21.97 15.4224 22.2805 15.4241C23.2469 15.4241 23.9798 15.0766 24.6891 14.7406C25.3553 14.4249 25.982 14.1283 26.7404 14.1283C26.9539 14.1288 27.1667 14.1522 27.3752 14.1985C29.6574 14.7 29.8151 17.0838 29.8165 17.1076C29.8275 17.292 29.6871 17.4503 29.5027 17.4613C29.5029 17.4613 29.5025 17.4613 29.5027 17.4613H29.4817C29.305 17.4608 29.1588 17.3235 29.1473 17.147C29.1423 17.0692 29.0062 15.2421 27.2312 14.8524C27.0704 14.8165 26.9061 14.7984 26.7414 14.7981C26.1305 14.7981 25.5953 15.0518 24.9766 15.3452C24.2365 15.6958 23.399 16.093 22.2771 16.093C21.9225 16.0915 21.569 16.0543 21.2218 15.9817C16.2833 14.9701 13.7307 10.5667 13.7056 10.5225C13.6142 10.3616 13.6703 10.1572 13.8312 10.0655C13.9922 9.97376 14.1966 10.0301 14.2883 10.1911L14.2895 10.193ZM22.7605 20.5921V18.8297L23.2054 18.6864C23.3546 18.6329 23.4883 18.5816 23.5994 18.5359L23.608 19.1683L22.7605 20.5921ZM22.7717 20.6363V23.5879L21.8979 25.1968V22.113L22.7717 20.6363ZM19.6481 25.8262L18.3117 28.0739V33.0377L17.4382 34.6466V29.543L18.3079 28.0715V19.1764L18.4846 19.2058C18.6995 19.2361 18.9107 19.2536 19.1194 19.2674L19.651 19.2868H19.6605V25.8048L20.5343 24.3273V19.256L20.8846 19.2287C21.0783 19.206 21.2608 19.1762 21.441 19.1449L21.8869 19.0556V22.0612L20.5338 24.3356V28.4398L19.66 30.0486L19.6605 25.8048L19.6481 25.8262ZM13.8549 35.5679V17.6736C14.0072 17.7811 14.1656 17.8797 14.3294 17.9693L14.8612 18.2319C14.9708 18.2797 15.0866 18.3294 15.2077 18.3774V33.2725L16.0813 31.7948V18.6821L16.088 18.6845C16.2878 18.7451 16.4939 18.8039 16.7136 18.8602L17.245 18.9818C17.3063 18.9949 17.3706 19.0071 17.4339 19.0198V29.5488L16.0813 31.8234V37.6467L15.2077 39.2558V33.2933L13.8549 35.5679ZM11.6278 39.3117V18.2167L11.9807 17.8632L12.8975 16.9407C12.9214 16.9539 12.9529 16.972 12.982 16.9885V37.037L11.9368 38.7949L11.6278 39.3117ZM11.6688 39.3203V46.1633L10.7953 47.7721V40.7978L11.6688 39.3203ZM12.9804 37.1001L13.8542 35.6224V42.4654L12.9804 44.0744V37.1001ZM9.03479 12.2482L13.4912 8.24791C13.5523 8.1925 13.6321 8.16193 13.7147 8.16193C13.9 8.16241 14.0497 8.31286 14.0495 8.49818C14.0495 8.50391 14.0493 8.50964 14.049 8.51537C14.0445 8.60421 14.0046 8.68755 13.938 8.74677L9.67336 12.5753C10.1715 13.5148 10.8939 14.3167 11.7768 14.9094C11.9339 15.0071 11.9819 15.2139 11.8842 15.371C11.7866 15.5282 11.5798 15.5762 11.4226 15.4785C10.3587 14.7697 9.50548 13.7875 8.9524 12.635C8.8927 12.5018 8.92613 12.3454 9.03479 12.2482ZM9.14153 14.9911C10.0767 14.9911 10.8347 15.7491 10.8347 16.6842C10.8347 17.6194 10.0767 18.3774 9.14153 18.3774C8.20637 18.3774 7.44839 17.6194 7.44839 16.6842C7.44935 15.7498 8.20684 14.9923 9.14153 14.9911ZM10.7547 19.1404V40.7811L10.7446 40.798L8.81079 44.0494C6.10654 38.9764 6.006 33.7236 6.00504 33.6596C5.70415 26.8426 8.37855 22.0295 10.7547 19.1404ZM5.8651 47.4509C5.86534 46.4556 6.67227 45.6487 7.66762 45.6489C8.66297 45.6491 9.46989 46.4561 9.46966 47.4514C9.46942 48.4468 8.66273 49.2535 7.66738 49.2535C6.67251 49.2525 5.86606 48.4461 5.8651 47.4509ZM11.2669 50.0188L8.30117 55.8851C8.24434 55.9981 8.12828 56.0692 8.00171 56.069C7.94941 56.069 7.89759 56.0566 7.85078 56.0332C7.68577 55.9503 7.61914 55.7495 7.70201 55.5845C7.70225 55.5838 7.70272 55.5833 7.70296 55.5826L10.6687 49.7165C10.7255 49.6035 10.8414 49.5321 10.9679 49.5324C11.1525 49.5317 11.3027 49.6807 11.3035 49.8653C11.3037 49.9188 11.291 49.9713 11.2669 50.0188ZM13.9072 50.2868C13.8596 50.4045 13.7565 50.491 13.6321 50.5175C13.5076 50.5437 13.3784 50.5065 13.287 50.4181C13.2817 50.4131 12.7399 49.893 11.1296 48.5726C10.9663 48.4389 10.9424 48.1982 11.0763 48.0348C11.2101 47.8715 11.4508 47.8476 11.6141 47.9816C12.5025 48.7092 13.0685 49.1957 13.4047 49.4925C14.9281 45.7244 24.3872 22.3155 24.9654 20.6585C25.4611 19.2373 26.1568 18.6544 26.6532 18.4158C27.0578 18.2095 27.529 18.1765 27.9583 18.3244C28.1573 18.3946 28.2616 18.6128 28.1914 18.8118C28.1212 19.0107 27.9029 19.1151 27.704 19.0448C27.4659 18.9668 27.2063 18.9883 26.9845 19.1045C26.467 19.3534 26.0056 19.9943 25.6868 20.9099C25.0544 22.7208 14.3619 49.1629 13.9072 50.2868ZM26.7655 23.8296C27.0401 23.8296 27.2629 24.0522 27.2629 24.327C27.2629 24.6016 27.0404 24.8245 26.7655 24.8245C26.4909 24.8245 26.2681 24.6019 26.2681 24.327C26.2678 24.0524 26.4909 23.8296 26.7655 23.8296C26.7652 23.8296 26.7655 23.8293 26.7655 23.8296ZM25.4463 27.4981C25.4478 27.683 25.2995 27.8341 25.1146 27.8365H25.112C24.9286 27.8356 24.78 27.688 24.7777 27.5046L24.7426 24.4963C24.7406 24.3113 24.8892 24.1599 25.0743 24.1579C25.2593 24.156 25.4107 24.3046 25.4126 24.4896L25.4463 27.4981ZM24.1826 27.9588C24.1823 28.2334 23.9597 28.4558 23.6851 28.4558C23.4105 28.4558 23.1882 28.2329 23.1882 27.9583C23.1884 27.6839 23.4107 27.4614 23.6854 27.4614C23.96 27.4609 24.1828 27.683 24.1835 27.9576C24.1835 27.9581 24.1826 27.9583 24.1826 27.9588ZM21.1471 37.9385C21.4217 37.9383 21.6445 38.1609 21.6448 38.4357C21.645 38.7106 21.4224 38.9332 21.1476 38.9334C20.8729 38.9336 20.6501 38.7111 20.6499 38.4362C20.6499 38.4365 20.6499 38.436 20.6499 38.4362C20.6496 38.1616 20.8724 37.9388 21.1471 37.9385ZM16.0648 47.2974C15.9497 47.4423 15.7386 47.4667 15.5936 47.3514C15.4487 47.2363 15.4243 47.0251 15.5397 46.8802L18.3855 43.2969C18.449 43.2167 18.5458 43.1698 18.6482 43.1701C18.8328 43.1689 18.9835 43.3174 18.9847 43.502C18.9852 43.5792 18.9591 43.6539 18.9109 43.7139L16.0648 47.2974ZM18.0703 47.5548C18.0703 47.8294 17.8477 48.052 17.5731 48.052C17.2985 48.052 17.0759 47.8294 17.0759 47.5548C17.0759 47.2802 17.2985 47.0576 17.5731 47.0576C17.8477 47.0576 18.0703 47.28 18.0703 47.5548ZM17.3813 42.1537C17.3811 41.8791 17.6037 41.6563 17.8785 41.6561C18.1532 41.6558 18.376 41.8784 18.3762 42.1532C18.3764 42.4279 18.1539 42.6507 17.879 42.6509C17.8792 42.6509 17.8788 42.6509 17.879 42.6509C17.6046 42.6507 17.3816 42.4281 17.3813 42.1537ZM16.7745 50.5253C16.4999 50.5256 16.2771 50.303 16.2768 50.0281C16.2766 49.7533 16.4992 49.5307 16.774 49.5305C17.0487 49.5302 17.2715 49.7528 17.2717 50.0277C17.2717 50.3023 17.0489 50.5251 16.7745 50.5253ZM18.9343 51.1971C18.9095 51.3507 18.7652 51.455 18.6117 51.4302C18.4751 51.4082 18.3748 51.2905 18.375 51.1522C18.3755 51.1214 18.381 51.0908 18.3912 51.0617C18.577 50.1062 17.9533 49.1813 16.9978 48.9955C16.0514 48.8114 15.133 49.4221 14.9362 50.3658C14.9333 50.3739 14.93 50.3821 14.9259 50.3897C14.8794 50.5378 14.7217 50.6201 14.5737 50.5736C14.4567 50.5368 14.3769 50.4284 14.3769 50.3059C14.3769 50.3009 14.3769 50.2961 14.3783 50.2911C14.3767 50.2729 14.3776 50.2543 14.3812 50.2364C14.6589 48.9815 15.9012 48.1893 17.1561 48.4668C18.3936 48.7407 19.1838 49.9544 18.9343 51.1971ZM19.3541 38.2657C19.5373 38.2688 19.6858 38.4145 19.6925 38.5976L19.7276 41.6463C19.73 41.8309 19.5822 41.9825 19.3976 41.9849C19.3971 41.9849 19.3966 41.9849 19.3961 41.9849H19.3926C19.2094 41.9837 19.0609 41.8364 19.0582 41.6532L19.0231 38.6043C19.0217 38.4197 19.1693 38.2683 19.3541 38.2657ZM19.0315 42.826C19.0317 42.5513 19.2543 42.329 19.5289 42.329C19.8035 42.3292 20.0259 42.5518 20.0259 42.8264C20.0256 43.1008 19.8033 43.3234 19.5287 43.3234C19.2543 43.3232 19.0315 43.1006 19.0315 42.826ZM19.9502 49.4392C19.9454 49.621 19.7976 49.7659 19.6158 49.7676H19.612C19.4262 49.7647 19.2777 49.6117 19.2806 49.4259L19.3904 43.9949C19.3966 43.8125 19.5459 43.6675 19.7286 43.6666C19.9144 43.669 20.0631 43.8213 20.0607 44.0071C20.0607 44.0076 20.0607 44.0081 20.0607 44.0083L19.9502 49.4392ZM21.1559 43.955C20.8813 43.955 20.6585 43.7325 20.6585 43.4576C20.6585 43.183 20.881 42.9602 21.1559 42.9602C21.4308 42.9602 21.6533 43.1827 21.6533 43.4574C21.6531 43.7318 21.4303 43.955 21.1559 43.955ZM23.8817 38.7113L20.4562 42.2263C20.327 42.3589 20.1147 42.3617 19.9822 42.2323C19.8496 42.1031 19.8468 41.8908 19.9762 41.7583L23.4024 38.243C23.5316 38.1105 23.7436 38.1076 23.8764 38.2368C24.0089 38.366 24.0118 38.5781 23.8826 38.7108L23.8817 38.7113ZM22.5448 37.0862C21.6166 37.0862 20.8641 36.3338 20.8641 35.4055C20.8641 34.4773 21.6166 33.7248 22.5448 33.7248C23.4731 33.7248 24.2255 34.4773 24.2255 35.4055C24.2241 36.333 23.4721 37.085 22.5448 37.0862ZM23.7802 32.8725C23.5055 32.8727 23.2827 32.6502 23.2825 32.3755C23.2823 32.1009 23.5048 31.8781 23.7797 31.8779C24.0543 31.8776 24.2771 32.1002 24.2774 32.3751C24.2771 32.6492 24.0546 32.8722 23.7802 32.8725ZM24.6306 29.5662L22.0959 32.7576C21.9808 32.9026 21.7699 32.9269 21.6247 32.8118C21.4797 32.6967 21.4554 32.4859 21.5705 32.3407L24.1052 29.149C24.1687 29.0688 24.2654 29.0222 24.3679 29.0224C24.5525 29.0212 24.7032 29.17 24.7041 29.3546C24.7046 29.4315 24.6786 29.5063 24.6306 29.5662ZM24.7141 28.5787C24.7139 28.3041 24.9365 28.0813 25.2113 28.0811C25.4862 28.0808 25.7088 28.3034 25.709 28.5783C25.7092 28.8531 25.4867 29.0757 25.2118 29.0759C25.212 29.0759 25.2116 29.0759 25.2118 29.0759C24.9372 29.0757 24.7139 28.8534 24.7141 28.5787ZM25.693 34.8904C25.6882 35.0721 25.5404 35.2171 25.3587 35.2188H25.3546C25.1688 35.2157 25.0205 35.0628 25.0232 34.877L25.1227 29.9538C25.1235 29.9282 25.1273 29.9029 25.1337 29.8781V29.8473C25.1366 29.753 25.2087 29.6751 25.3026 29.6648C25.3508 29.6395 25.4045 29.6259 25.459 29.6257C25.645 29.6269 25.7945 29.7787 25.7933 29.9645C25.7933 29.9655 25.7933 29.9662 25.7933 29.9672L25.693 34.8904ZM26.7058 29.6737C26.4312 29.6739 26.2084 29.4513 26.2081 29.1765C26.2079 28.9016 26.4304 28.679 26.7053 28.6788C26.9799 28.6786 27.2027 28.9011 27.2027 29.176C27.2023 29.4501 26.9799 29.6732 26.7058 29.6737ZM29.6142 24.6707L26.1881 28.1857C26.0596 28.3182 25.8478 28.3213 25.7155 28.1926C25.7152 28.1923 25.7147 28.1919 25.7145 28.1916C25.6247 28.1061 25.5913 27.9774 25.6278 27.859C25.6218 27.8038 25.6405 27.7486 25.6789 27.7085L29.1051 24.1935C29.1364 24.1622 29.1777 24.1431 29.2219 24.1396C29.3857 24.0543 29.5875 24.1181 29.6727 24.2819C29.7394 24.4104 29.7157 24.5673 29.6142 24.6707ZM28.312 23.341C27.3673 23.3412 26.6014 22.5754 26.6014 21.6309C26.6012 20.6862 27.367 19.9203 28.3115 19.9203C29.2562 19.9201 30.0221 20.6859 30.0221 21.6304C30.0221 21.6302 30.0221 21.6306 30.0221 21.6304C30.0207 22.5742 29.2555 23.3398 28.312 23.341Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M43.7011 0.0019349C43.5038 10.3978 43.3984 21.0838 43.3984 32C43.3984 42.9162 43.5038 53.6022 43.7011 63.9981C43.7862 63.9981 43.8704 64 43.9555 64C44.0406 64 44.1247 63.9981 44.2098 63.9981C44.4071 53.6022 44.5125 42.9162 44.5125 32C44.5125 21.0838 44.4071 10.3978 44.2098 0.0019349C44.1247 0.0019349 44.0406 0 43.9555 0C43.8704 0 43.7862 0.0019349 43.7011 0.0019349Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M65.0935 31.1514C64.2424 31.1514 63.1754 30.7186 62.5803 30.1379C61.7693 29.3546 61.5668 29.0164 60.5266 26.7334C60.3241 26.7602 60.175 26.7736 60.0805 26.7736C59.9859 26.7736 59.8636 26.7602 59.6755 26.7334V28.5579C59.6755 29.6125 59.878 29.8551 60.8915 30.0443V30.328H56.5137V30.0443C57.6485 29.9086 57.9054 29.6259 57.9054 28.5579V23.6127C57.9054 22.5725 57.6752 22.2343 56.7028 21.8971V21.6134L59.3106 21.7625C60.1483 21.6134 60.5533 21.5733 61.0663 21.5733C62.6338 21.5733 63.5661 22.4244 63.5661 23.8296C63.5661 24.2212 63.4983 24.5861 63.3493 24.8841C63.1066 25.3837 62.7818 25.7085 61.9164 26.316C63.2681 28.4911 63.93 29.4234 64.5509 30.0051C64.997 30.4235 65.51 30.6537 66.1996 30.7483V30.9775C65.6341 31.1246 65.445 31.1514 65.0935 31.1514ZM59.6755 22.1799V25.9492C60.0805 26.0982 60.2562 26.1383 60.5266 26.1383C60.7023 26.1383 60.8236 26.125 61.1752 26.0438C61.2697 25.9358 61.3509 25.8269 61.4187 25.7467C61.7292 25.3283 61.8916 24.8278 61.8916 24.2604C61.8916 23.0043 61.1885 22.3413 59.6755 22.1799Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M65.5701 30.327V30.0433C65.7459 29.9888 65.8806 29.9354 65.9885 29.8675C66.2321 29.6918 66.2856 29.5294 66.3124 28.6515L66.4069 26.0438V25.9492C66.4069 25.6111 66.3525 25.2605 66.2856 25.0714C66.2054 24.8555 66.0697 24.7609 65.7459 24.6386V24.3683L67.6105 23.99L68.0833 24.1792L67.9888 25.8546V28.6515C67.9888 29.7185 68.0556 29.8675 68.6374 30.0433V30.327H65.5701ZM67.2456 22.6795C66.6247 22.6795 66.1242 22.2334 66.1242 21.666C66.1242 21.1521 66.6238 20.7337 67.2456 20.7337C67.8541 20.7337 68.3537 21.1521 68.3537 21.666C68.3537 22.2343 67.8541 22.6795 67.2456 22.6795Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M73.4348 30.4628H73.408C72.9485 30.4628 72.705 30.3004 72.5158 29.8676C71.786 30.3816 71.4211 30.5162 70.8002 30.5162C69.8679 30.5162 69.2461 29.9622 69.2461 29.1379C69.2461 28.8265 69.3273 28.5428 69.4763 28.3135C69.6654 28.0298 69.9358 27.8674 70.6512 27.6238L72.5158 26.9752V26.1375L72.4891 25.3943C72.4757 24.868 72.3811 24.7324 71.9895 24.7324C71.53 24.7324 71.2597 24.9616 70.976 25.5968L70.6522 26.3133H70.2739C69.8145 26.3133 69.5308 26.1241 69.5308 25.786C69.5308 25.4889 69.7734 25.124 70.1794 24.8269C70.8958 24.2996 71.7469 24.0303 72.7193 24.0303C73.706 24.0303 74.0977 24.3952 74.0977 25.3141V25.4755L74.0031 28.4205V28.515C74.0031 29.2582 74.1655 29.5018 74.9354 29.8667V30.2316C74.3537 30.3682 73.8541 30.4628 73.4348 30.4628ZM72.5158 27.5847L71.5835 28.003C71.0429 28.2466 70.827 28.4893 70.827 28.8952C70.827 29.3413 71.1508 29.625 71.638 29.625C71.8672 29.625 72.1653 29.5849 72.5158 29.4894V27.5847Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M75.3555 30.327V30.0433C75.9362 29.922 76.0585 29.6918 76.0986 28.6515L76.1932 26.0438V23.5306C76.1932 22.4359 76.1664 22.1789 76.0585 21.815C75.9907 21.5714 75.9095 21.4635 75.6258 21.2877V21.004L77.6384 20.6391L77.9623 20.9094C77.7865 22.071 77.7731 22.3681 77.7731 24.2193V28.6515C77.7731 29.7185 77.9489 29.9487 78.4895 30.0433V30.327H75.3555Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M81.3558 30.517C80.2478 30.517 79.6259 29.9629 79.6536 28.9361L79.7482 25.4772L79.0996 24.6395V24.2746C79.6536 24.0577 79.8561 23.9364 80.2478 23.5448C80.5047 23.3012 80.6127 23.1665 80.8696 22.7739L81.5182 23.0576L81.3291 24.3137L82.4505 24.2192C82.7887 24.1924 83.0046 24.4083 83.0046 24.692C83.0046 25.0435 82.7075 25.3005 82.2614 25.3005H81.3291V28.4623C81.3291 29.0297 81.4237 29.2866 81.6806 29.4624C81.9242 29.6248 82.1945 29.6792 82.91 29.6792L83.0046 29.9496C82.3015 30.3947 81.9643 30.517 81.3558 30.517Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M87.5833 30.4626H87.5565C87.0971 30.4626 86.8535 30.3003 86.6644 29.8675C85.9346 30.3814 85.5697 30.5161 84.9488 30.5161C84.0165 30.5161 83.3946 29.9621 83.3946 29.1377C83.3946 28.8263 83.4758 28.5426 83.6248 28.3134C83.814 28.0297 84.0843 27.8673 84.7998 27.6237L86.6644 26.9751V26.1374L86.6376 25.3942C86.6242 24.8679 86.5297 24.7323 86.138 24.7323C85.6786 24.7323 85.4082 24.9615 85.1245 25.5967L84.8007 26.3131H84.4224C83.963 26.3131 83.6793 26.124 83.6793 25.7859C83.6793 25.4888 83.9219 25.1239 84.3279 24.8268C85.0443 24.2995 85.8954 24.0302 86.8678 24.0302C87.8546 24.0302 88.2462 24.3951 88.2462 25.314V25.4754L88.1516 28.4204V28.5149C88.1516 29.2581 88.314 29.5017 89.0839 29.8666V30.2315C88.5013 30.3681 88.0017 30.4626 87.5833 30.4626ZM86.6644 27.5846L85.7321 28.0029C85.1914 28.2465 84.9755 28.4891 84.9755 28.8951C84.9755 29.3412 85.2993 29.6249 85.7865 29.6249C86.0158 29.6249 86.3138 29.5848 86.6644 29.4893V27.5846Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M92.2467 30.4626C91.6927 30.4626 91.2876 30.4225 90.0038 30.1923L89.8147 28.3688L90.1796 28.2742C90.8282 29.0575 91.0985 29.3412 91.45 29.5982C91.6525 29.7596 91.9773 29.8819 92.1798 29.8819C92.6125 29.8819 92.9497 29.5303 92.9497 29.0986C92.9497 28.9773 92.923 28.855 92.8695 28.7604C92.815 28.6391 92.7205 28.5436 92.6125 28.4767C92.518 28.4223 92.3288 28.3153 92.0451 28.1797L91.302 27.8148C90.2073 27.2741 89.7755 26.7334 89.7755 25.9091C89.7755 24.7743 90.6534 24.0311 92.005 24.0311C92.6125 24.0311 92.9641 24.098 93.8152 24.3693L94.0043 25.761L93.626 25.9502C93.0854 25.2882 92.9096 25.0991 92.6125 24.8832C92.4234 24.7342 92.113 24.6262 91.8961 24.6262C91.45 24.6262 91.1128 24.9644 91.1128 25.3828C91.1128 25.761 91.3555 26.018 91.9506 26.3151L93.072 26.8825C94.031 27.3687 94.3826 27.8014 94.3826 28.5445C94.3816 29.7051 93.5295 30.4626 92.2467 30.4626Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M101.733 30.3269V30.0432C102.328 29.9219 102.45 29.7051 102.49 28.6515L102.584 26.314V26.2462C102.584 25.4219 102.179 24.9901 101.435 24.9901C101.125 24.9901 100.908 25.0579 100.436 25.3015V28.6524C100.436 29.7194 100.516 29.9229 100.99 30.0442V30.3279H98.0176V30.0442C98.5984 29.9229 98.7197 29.6927 98.7607 28.6524L98.8553 26.0447V25.9234C98.8553 25.1401 98.7072 24.8831 98.0988 24.6395V24.3683L100.436 23.99V24.8277C101.301 24.1791 101.72 24.0167 102.571 24.0167C103.679 24.0167 104.165 24.5163 104.165 25.6655V28.6515C104.165 29.6249 104.273 29.8675 104.814 30.0432V30.3269H101.733Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M109.614 30.4628H109.588C109.128 30.4628 108.885 30.3004 108.695 29.8676C107.966 30.3816 107.601 30.5162 106.98 30.5162C106.048 30.5162 105.426 29.9622 105.426 29.1379C105.426 28.8265 105.507 28.5428 105.656 28.3135C105.845 28.0298 106.115 27.8674 106.831 27.6238L108.695 26.9752V26.1375L108.669 25.3943C108.655 24.868 108.561 24.7324 108.169 24.7324C107.71 24.7324 107.439 24.9616 107.156 25.5968L106.832 26.3133H106.454C105.994 26.3133 105.71 26.1241 105.71 25.786C105.71 25.4889 105.953 25.124 106.359 24.8269C107.075 24.2996 107.927 24.0303 108.899 24.0303C109.886 24.0303 110.277 24.3952 110.277 25.3141V25.4755L110.183 28.4205V28.515C110.183 29.2582 110.345 29.5018 111.115 29.8667V30.2316C110.533 30.3682 110.034 30.4628 109.614 30.4628ZM108.695 27.5847L107.763 28.003C107.223 28.2466 107.007 28.4893 107.007 28.8952C107.007 29.3413 107.33 29.625 107.818 29.625C108.047 29.625 108.345 29.5849 108.695 29.4894V27.5847Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M118.061 30.327V30.0433C118.642 29.8408 118.763 29.6249 118.804 28.6516L118.899 26.4087V26.3008C118.899 25.3818 118.588 25.017 117.831 25.017C117.507 25.017 117.277 25.0848 116.751 25.3007V28.6516C116.751 29.6106 116.859 29.8676 117.305 30.0433V30.327H114.332V30.0433C114.967 29.9077 115.035 29.773 115.075 28.6516L115.17 26.0438V22.5439C115.17 21.774 115.008 21.517 114.427 21.2878V21.0041L116.562 20.6392L116.94 21.0041L116.751 24.8278C117.656 24.1658 118.116 23.9767 118.859 23.9767C119.345 23.9767 119.751 24.1114 120.048 24.355C120.359 24.5976 120.481 24.9625 120.481 25.6655V28.6516C120.481 29.7319 120.549 29.8943 121.129 30.0433V30.327H118.061Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M127.846 30.327H121.414V30.0433C122.535 29.8943 122.806 29.6106 122.806 28.557V23.6118C122.806 22.4903 122.643 22.2066 121.792 21.9363V21.666H127.656L127.846 23.7063L127.562 23.8009C126.967 22.7339 126.522 22.3958 125.603 22.3146L124.576 22.22V25.5709L125.508 25.5308C126.467 25.4907 126.67 25.3283 126.913 24.4638H127.197V27.2474H126.913C126.643 26.423 126.386 26.2205 125.508 26.1804L124.576 26.1403V28.3697C124.576 29.3555 124.86 29.6803 125.724 29.6803C126.724 29.6803 127.318 29.2342 128.129 27.9103L128.4 28.086L127.846 30.327ZM126.386 20.5455L123.779 21.005L123.685 20.8292L126.022 19.4241L126.671 19.8969L126.386 20.5455Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M129.16 30.327V30.0433C129.336 29.9888 129.471 29.9353 129.579 29.8675C129.822 29.6917 129.876 29.5293 129.902 28.6515L129.997 26.0437V25.9492C129.997 25.611 129.942 25.2604 129.876 25.0713C129.795 24.8554 129.66 24.7609 129.336 24.6386V24.3683L131.201 23.99L131.673 24.1791L131.579 25.8546V28.6515C131.579 29.7185 131.646 29.8675 132.227 30.0433V30.327H129.16ZM130.836 22.6794C130.215 22.6794 129.714 22.2333 129.714 21.6659C129.714 21.152 130.214 20.7336 130.836 20.7336C131.444 20.7336 131.944 21.152 131.944 21.6659C131.944 22.2343 131.444 22.6794 130.836 22.6794Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M137.077 26.2328C136.659 25.8947 136.253 25.7065 135.943 25.7065C135.727 25.7065 135.538 25.8545 135.307 26.1793V28.6514C135.307 29.6105 135.51 29.8407 136.523 30.0432V30.3269H132.889V30.0432C133.497 29.8264 133.605 29.6382 133.632 28.6514L133.726 25.8545V25.7867C133.726 25.1381 133.537 24.8687 132.97 24.6385V24.3682L135.307 23.9766L135.253 25.3817H135.307C135.847 24.3959 136.28 23.9899 136.807 23.9899C137.225 23.9899 137.509 24.3949 137.509 24.9624C137.51 25.3951 137.429 25.652 137.077 26.2328Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M143.013 26.9762H139.472V27.1242C139.472 27.7194 139.486 27.8139 139.567 28.1788C139.77 29.0032 140.432 29.4894 141.445 29.4894C141.959 29.4894 142.256 29.4216 142.918 29.1245L143.107 29.2191C143.026 29.4216 142.971 29.5839 142.918 29.6374C142.256 30.286 141.405 30.5697 140.458 30.5697C138.959 30.5697 137.891 29.2859 137.891 27.5159C137.891 25.4622 139.12 23.9758 140.85 23.9758C142.228 23.9758 143.134 24.7324 143.134 25.8939C143.134 26.1662 143.094 26.5311 143.013 26.9762ZM141.486 25.0715C141.365 24.8011 141.094 24.6388 140.756 24.6388C140.405 24.6388 140.081 24.8011 139.838 25.1116C139.621 25.3953 139.54 25.6389 139.473 26.233H141.433L141.608 26.0439C141.607 25.4631 141.581 25.3017 141.486 25.0715Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M147.917 30.4628H147.89C147.431 30.4628 147.187 30.3004 146.998 29.8676C146.268 30.3816 145.904 30.5162 145.283 30.5162C144.35 30.5162 143.729 29.9622 143.729 29.1379C143.729 28.8265 143.81 28.5428 143.959 28.3135C144.148 28.0298 144.418 27.8674 145.134 27.6238L146.998 26.9752V26.1375L146.972 25.3943C146.958 24.868 146.864 24.7324 146.472 24.7324C146.012 24.7324 145.742 24.9616 145.458 25.5968L145.135 26.3133H144.756C144.297 26.3133 144.013 26.1241 144.013 25.786C144.013 25.4889 144.256 25.124 144.662 24.8269C145.378 24.2996 146.229 24.0303 147.202 24.0303C148.188 24.0303 148.58 24.3952 148.58 25.3141V25.4755L148.486 28.4205V28.515C148.486 29.2582 148.648 29.5018 149.418 29.8667V30.2316C148.836 30.3682 148.337 30.4628 147.917 30.4628ZM146.998 27.5847L146.066 28.003C145.525 28.2466 145.309 28.4893 145.309 28.8952C145.309 29.3413 145.633 29.625 146.12 29.625C146.35 29.625 146.647 29.5849 146.998 29.4894V27.5847Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M153.649 30.327V30.0433C154.244 29.922 154.366 29.7051 154.406 28.6515L154.501 26.3141V26.2463C154.501 25.4219 154.096 24.9902 153.351 24.9902C153.041 24.9902 152.824 25.058 152.352 25.3016V28.6525C152.352 29.7195 152.432 29.9229 152.906 30.0443V30.328H149.934V30.0443C150.514 29.9229 150.637 29.6927 150.677 28.6525L150.771 26.0447V25.9234C150.771 25.1401 150.623 24.8832 150.015 24.6396V24.3683L152.352 23.99V24.8278C153.217 24.1792 153.636 24.0168 154.487 24.0168C155.595 24.0168 156.081 24.5164 156.081 25.6655V28.6515C156.081 29.6249 156.189 29.8675 156.73 30.0433V30.327H153.649Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M161.11 30.3269V30.0432C161.705 29.9219 161.827 29.7051 161.867 28.6515L161.962 26.314V26.2462C161.962 25.4219 161.556 24.9901 160.813 24.9901C160.503 24.9901 160.286 25.0579 159.814 25.3015V28.6524C159.814 29.7194 159.894 29.9229 160.367 30.0442V30.3279H157.395V30.0442C157.975 29.9229 158.098 29.6927 158.138 28.6524L158.232 26.0447V25.9234C158.232 25.1401 158.084 24.8831 157.476 24.6395V24.3683L159.814 23.99V24.8277C160.679 24.1791 161.098 24.0167 161.948 24.0167C163.056 24.0167 163.542 24.5163 163.542 25.6655V28.6515C163.542 29.6249 163.65 29.8675 164.191 30.0432V30.3269H161.11Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M64.6212 41.6845V42.6168L64.8103 43.914H64.3919C64.2295 43.914 64.1617 43.9274 63.1358 44.2522C62.6362 44.4012 62.0277 44.4824 61.3524 44.4824C59.974 44.4824 59.163 44.1987 58.3797 43.4555C57.4742 42.591 56.9746 41.3884 56.9746 40.0368C56.9746 38.7663 57.4207 37.5646 58.1906 36.7393C59.0283 35.8481 60.0686 35.4421 61.5138 35.4421C62.4194 35.4421 63.0002 35.5501 64.3384 35.9962L64.433 38.0499L64.1627 38.239C63.3116 36.5502 62.5951 35.9962 61.2712 35.9962C59.5413 35.9962 58.3797 37.5771 58.3797 39.9145C58.3797 42.2252 59.7037 43.8328 61.6084 43.8328C62.8645 43.8328 63.4997 43.3189 63.4997 42.3331V41.7791C63.4997 40.82 63.2428 40.5363 62.2025 40.3329V40.0091H65.648V40.2794C64.8103 40.5497 64.6212 40.8067 64.6212 41.6845Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M69.3675 44.4824C67.7054 44.4824 66.7598 43.4145 66.7598 41.5365C66.7598 39.2937 68.0159 37.9153 70.0295 37.9153C71.6916 37.9153 72.6373 38.9689 72.6373 40.8602C72.6373 43.1041 71.3945 44.4824 69.3675 44.4824ZM69.6646 38.4693C68.6779 38.4693 67.9758 39.5506 67.9758 41.0771C67.9758 42.7392 68.7056 43.9284 69.7324 43.9284C70.7192 43.9284 71.4213 42.8204 71.4213 41.2805C71.4213 39.6452 70.6915 38.4693 69.6646 38.4693Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M78.5979 38.5094C78.4498 38.6174 78.2597 38.9011 78.1394 39.1714L76.0856 43.9141C75.8965 44.3468 75.7074 44.5761 75.5182 44.5761H75.4781L73.7482 39.1714C73.6269 38.7931 73.4368 38.5629 73.0996 38.4282V38.1445H75.5182V38.4282L75.2479 38.5085C74.9375 38.603 74.8295 38.7387 74.8295 38.9947C74.8295 39.1303 74.8696 39.3462 74.9652 39.6299L75.992 42.8051L77.1135 39.819C77.2348 39.4952 77.2892 39.2249 77.2892 39.0224C77.2892 38.6842 77.0189 38.4951 76.4515 38.4683V38.1445H78.7889V38.4282C78.6924 38.4559 78.6389 38.4961 78.5979 38.5094Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M81.6256 44.4415C80.2205 44.4415 79.248 43.2112 79.248 41.4287C79.248 39.3072 80.4096 37.9966 82.2608 37.9966C83.4902 37.9966 84.1923 38.6175 84.1923 39.6854C84.1923 39.9557 84.1655 40.3063 84.0977 40.6989H80.274V41.0504C80.274 42.5769 81.0171 43.4824 82.2742 43.4824C82.7738 43.4824 83.0852 43.4012 83.963 43.0774L84.2334 43.1319C84.0719 43.8063 82.8148 44.4415 81.6256 44.4415ZM83.1664 39.631C83.1664 39.0636 82.6668 38.6041 82.0172 38.6041C81.2607 38.6041 80.6799 39.1448 80.3685 40.1449H82.9763L83.1654 39.9147L83.1664 39.631Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M89.0995 39.4283L88.9916 39.604C88.6267 39.4149 88.2895 39.3203 88.0182 39.3203C87.6533 39.3203 87.3562 39.5907 87.0314 40.1714V42.5222C87.0314 43.5902 87.3295 43.8739 88.5178 43.9685V44.2923H85.0723V43.9685C85.7476 43.7927 85.8832 43.5768 85.9234 42.5222L86.0046 40.2794C86.0179 40.1982 86.0179 40.118 86.0179 40.0368C86.0179 39.1178 85.91 38.9153 85.2614 38.6039V38.3336L87.0314 37.9553V39.4951C88.0048 38.239 88.3019 38.0222 89.004 38.0222C89.2466 38.0222 89.4367 38.2113 89.4367 38.4549C89.4376 38.7396 89.3154 39.0768 89.0995 39.4283Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M93.535 44.2924V43.9685C94.2782 43.8472 94.4262 43.6304 94.4673 42.5223L94.5619 40.2795V40.1983C94.5619 39.3605 94.1435 38.8476 93.4538 38.8476C92.9944 38.8476 92.6705 38.9421 91.8596 39.3472V42.5223C91.8596 43.0085 91.8997 43.3467 91.9541 43.4957C92.0487 43.7383 92.3056 43.874 92.7919 43.9685V44.2924H89.9004V43.9685C90.5891 43.7928 90.7114 43.5903 90.7515 42.5223L90.8327 40.2795C90.8461 40.225 90.8461 40.1581 90.8461 40.0903C90.8461 39.1981 90.6837 38.8743 90.0895 38.604V38.3337L91.7249 37.9153L91.8596 37.9965V38.9288C93.0354 38.1856 93.5751 37.9564 94.2247 37.9564C94.5762 37.9564 95.008 38.1178 95.2248 38.3346C95.4674 38.5773 95.5897 39.01 95.5897 39.6318V42.5233C95.5897 43.618 95.6709 43.7393 96.4274 43.9695V44.2933L93.535 44.2924Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M104.629 44.2924V43.9686C105.372 43.8472 105.521 43.6304 105.562 42.5223L105.656 40.2795V40.1849C105.656 39.3605 105.238 38.8333 104.588 38.8333C104.142 38.8333 103.832 38.968 102.954 39.4952V42.5214C102.954 43.0076 102.994 43.3457 103.048 43.4948C103.143 43.7374 103.4 43.873 103.886 43.9676V44.2914H100.9V43.9676C101.657 43.8463 101.791 43.6294 101.832 42.5214L101.927 40.2785V40.184C101.927 39.3596 101.522 38.8323 100.86 38.8323C100.455 38.8323 100.251 38.8992 99.2246 39.3453V42.5204C99.2246 43.0066 99.2647 43.3448 99.3192 43.4938C99.4137 43.7364 99.6707 43.8721 100.157 43.9666V44.2905H97.2654V43.9666C97.9541 43.7909 98.0764 43.5884 98.1165 42.5204L98.1977 40.2776C98.2111 40.1964 98.2111 40.1161 98.2111 40.0349C98.2111 39.116 98.1032 38.9135 97.4546 38.6021V38.3318L99.13 37.9535L99.2246 38.0481V38.9259C100.387 38.1828 100.954 37.9535 101.657 37.9535C102.373 37.9535 102.643 38.1961 102.954 39.0749C104.007 38.2105 104.561 37.9535 105.331 37.9535C105.764 37.9535 106.156 38.1426 106.399 38.4263C106.588 38.669 106.683 39.0339 106.683 39.629V42.5204C106.683 43.6151 106.764 43.7498 107.521 43.9666V44.2905H104.629V44.2924Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M110.835 44.4415C109.429 44.4415 108.457 43.2112 108.457 41.4287C108.457 39.3072 109.619 37.9966 111.47 37.9966C112.699 37.9966 113.401 38.6175 113.401 39.6854C113.401 39.9557 113.375 40.3063 113.307 40.6989H109.483V41.0504C109.483 42.5769 110.226 43.4824 111.483 43.4824C111.983 43.4824 112.294 43.4012 113.172 43.0774L113.442 43.1319C113.28 43.8063 112.023 44.4415 110.835 44.4415ZM112.374 39.631C112.374 39.0636 111.875 38.6041 111.225 38.6041C110.469 38.6041 109.888 39.1448 109.577 40.1449H112.184L112.373 39.9147L112.374 39.631Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M117.918 44.2924V43.9685C118.661 43.8472 118.809 43.6304 118.85 42.5223L118.945 40.2795V40.1983C118.945 39.3605 118.526 38.8476 117.837 38.8476C117.377 38.8476 117.053 38.9421 116.242 39.3472V42.5223C116.242 43.0085 116.282 43.3467 116.337 43.4957C116.431 43.7383 116.688 43.874 117.175 43.9685V44.2924H114.283V43.9685C114.972 43.7928 115.094 43.5903 115.134 42.5223L115.215 40.2795C115.229 40.225 115.229 40.1581 115.229 40.0903C115.229 39.1981 115.066 38.8743 114.472 38.604V38.3337L116.108 37.9153L116.242 37.9965V38.9288C117.418 38.1856 117.958 37.9564 118.607 37.9564C118.959 37.9564 119.391 38.1178 119.608 38.3346C119.85 38.5773 119.973 39.01 119.973 39.6318V42.5233C119.973 43.618 120.054 43.7393 120.81 43.9695V44.2933L117.918 44.2924Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M123.715 44.4279C122.675 44.4279 122.228 44.0095 122.256 42.9149L122.336 39.2267L121.701 38.5103L121.741 38.1454C122.12 37.9429 122.268 37.835 122.484 37.6315C122.741 37.3746 122.849 37.2131 123.12 36.7537L123.485 36.8893L123.377 38.2944L125.052 38.1865C125.296 38.1731 125.43 38.3078 125.43 38.5246C125.43 38.8227 125.16 39.0242 124.782 39.0242H123.377V42.335C123.377 43.2396 123.687 43.7402 124.254 43.7402C124.565 43.7402 124.971 43.6189 125.429 43.402L125.524 43.6857C124.864 44.3066 124.553 44.4279 123.715 44.4279Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M131.204 44.4825C129.541 44.4825 128.596 43.4145 128.596 41.5366C128.596 39.2937 129.852 37.9153 131.866 37.9153C133.528 37.9153 134.473 38.9689 134.473 40.8603C134.473 43.1041 133.231 44.4825 131.204 44.4825ZM131.501 38.4693C130.514 38.4693 129.812 39.5506 129.812 41.0771C129.812 42.7392 130.542 43.9284 131.568 43.9284C132.555 43.9284 133.257 42.8204 133.257 41.2806C133.257 39.6452 132.528 38.4693 131.501 38.4693Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M140.543 35.9427C139.936 35.3619 139.544 35.1594 139.03 35.1594C138.503 35.1594 138.138 35.4164 137.828 35.9972C137.611 36.4155 137.557 36.7671 137.49 37.6994L137.449 38.2801L139.125 38.1856C139.367 38.1722 139.503 38.3212 139.503 38.5371C139.503 38.902 139.409 38.9832 138.949 38.9832H137.449L137.544 42.5232C137.557 42.9416 137.584 43.2253 137.639 43.361C137.76 43.6848 138.138 43.8749 138.855 43.9695V44.2933H135.49V43.9695C136.233 43.8481 136.423 43.5367 136.423 42.5232V39.3347L135.585 38.6995V38.5103L136.49 37.9429C136.571 36.5779 136.774 36.0105 137.409 35.3753C137.909 34.8614 138.639 34.5911 139.517 34.5911C140.626 34.5911 141.274 34.8213 141.274 35.2129C141.273 35.5644 141.043 35.7803 140.543 35.9427Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M142.41 44.2924V44.0087C143.491 43.8463 143.707 43.5903 143.707 42.5223V37.4014C143.707 36.3611 143.505 36.1042 142.505 35.9017V35.6313H146.139V35.9017C145.058 36.0774 144.829 36.3344 144.829 37.4014V42.5223C144.829 43.5903 145.058 43.8463 146.139 44.0087V44.2924H142.41Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M150.896 39.4284L150.789 39.6041C150.424 39.415 150.086 39.3204 149.815 39.3204C149.45 39.3204 149.153 39.5908 148.828 40.1715V42.5223C148.828 43.5903 149.126 43.874 150.315 43.9686V44.2924H146.869V43.9686C147.545 43.7928 147.68 43.5769 147.72 42.5223L147.802 40.2795C147.815 40.1983 147.815 40.118 147.815 40.0368C147.815 39.1179 147.707 38.9154 147.058 38.604V38.3337L148.828 37.9554V39.4952C149.802 38.2391 150.099 38.0223 150.801 38.0223C151.044 38.0223 151.234 38.2114 151.234 38.455C151.234 38.7397 151.112 39.0768 150.896 39.4284Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M154.167 44.4415C152.761 44.4415 151.789 43.2112 151.789 41.4287C151.789 39.3072 152.951 37.9966 154.802 37.9966C156.031 37.9966 156.733 38.6175 156.733 39.6854C156.733 39.9557 156.707 40.3063 156.639 40.6989H152.815V41.0504C152.815 42.5769 153.558 43.4824 154.815 43.4824C155.315 43.4824 155.626 43.4012 156.504 43.0774L156.774 43.1319C156.613 43.8063 155.356 44.4415 154.167 44.4415ZM155.707 39.631C155.707 39.0636 155.208 38.6041 154.558 38.6041C153.802 38.6041 153.221 39.1448 152.91 40.1449H155.517L155.706 39.9147L155.707 39.631Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M157.48 44.2923V43.9685C158.224 43.8472 158.373 43.617 158.413 42.5223L158.507 39.9145V37.3211C158.507 35.916 158.399 35.6055 157.859 35.3619V34.9703L159.534 34.7L159.723 34.8757L159.534 38.3346V42.5232C159.534 43.0094 159.561 43.3476 159.63 43.4966C159.724 43.7392 159.968 43.8749 160.467 43.9694V44.2933L157.48 44.2923Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M165.98 44.3468C165.778 44.3736 165.575 44.4013 165.413 44.4013C164.859 44.4013 164.683 44.2523 164.521 43.7116C163.71 44.2924 163.332 44.428 162.616 44.428C161.724 44.428 161.17 43.9418 161.17 43.1442C161.17 42.4679 161.535 42.0896 162.562 41.7114L164.614 40.9548V39.82C164.614 38.9422 164.399 38.6852 163.655 38.6852C162.98 38.6852 162.71 38.9278 162.372 39.82C162.331 39.928 162.277 40.077 162.182 40.2795C161.683 40.2117 161.439 40.0368 161.439 39.7398C161.439 39.2803 162.182 38.6183 163.101 38.2668C163.507 38.1178 164.101 37.9965 164.466 37.9965C165.318 37.9965 165.682 38.3748 165.682 39.2392V39.4962L165.547 42.658V42.7659C165.547 43.4279 165.736 43.6438 166.654 43.9686V44.2656L165.98 44.3468ZM163.238 41.9817C162.617 42.2386 162.319 42.5625 162.319 42.9818C162.319 43.3734 162.684 43.7383 163.09 43.7383C163.333 43.7383 163.684 43.6571 164.116 43.4957C164.211 43.469 164.346 43.4145 164.522 43.3601L164.616 41.4143L163.238 41.9817Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M170.941 44.2924V43.9685C171.684 43.8472 171.833 43.6304 171.874 42.5223L171.969 40.2795V40.1983C171.969 39.3605 171.551 38.8476 170.861 38.8476C170.403 38.8476 170.078 38.9421 169.267 39.3472V42.5223C169.267 43.0085 169.307 43.3467 169.361 43.4957C169.457 43.7383 169.713 43.874 170.199 43.9685V44.2924H167.309V43.9685C167.997 43.7928 168.12 43.5903 168.16 42.5223L168.241 40.2795C168.255 40.225 168.255 40.1581 168.255 40.0903C168.255 39.1981 168.093 38.8743 167.498 38.604V38.3337L169.133 37.9153L169.269 37.9965V38.9288C170.444 38.1856 170.984 37.9564 171.634 37.9564C171.984 37.9564 172.417 38.1178 172.634 38.3346C172.877 38.5773 172.999 39.01 172.999 39.6318V42.5233C172.999 43.618 173.08 43.7393 173.837 43.9695V44.2933L170.941 44.2924Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M179.238 44.4824L179.062 44.2932V43.6446C178.36 44.2388 177.819 44.4824 177.131 44.4824C175.739 44.4824 174.766 43.2521 174.766 41.482C174.766 39.4694 176.049 37.9563 177.766 37.9563C178.198 37.9563 178.455 38.0107 179.062 38.24V37.2255C179.062 35.8481 178.845 35.4555 177.981 35.253V34.9292L180.075 34.699L180.264 34.8747L180.075 38.1445V42.5223C180.075 43.36 180.224 43.5768 180.927 43.7784V44.1022L179.238 44.4824ZM179.063 40.0091C179.063 39.1446 178.577 38.5629 177.848 38.5629C176.753 38.5629 175.888 39.6308 175.888 40.9815C175.888 42.5347 176.712 43.5892 177.928 43.5892C178.347 43.5892 178.564 43.5348 179.063 43.2645V40.0091Z"
+          fill="#0B0C0C"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_209_1004">
+          <rect width="180.926" height="64" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export default HarpBlack;
