+++ packages/react/ds/src/assets/logos/harp/harp-black.tsx
@@ -1,0 +1,45 @@
+function HarpBlack() {
+  return (
+    <svg
+      width="56"
+      height="88"
+      viewBox="0 0 56 88"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_212_1062)">
+        <path
+          d="M16.3935 18.5387C17.3014 18.5387 18.0376 17.8024 18.0376 16.8946C18.0376 15.9867 17.3014 15.2505 16.3935 15.2505C15.4857 15.2505 14.7495 15.9867 14.7495 16.8946C14.7503 17.802 15.4857 18.5378 16.3935 18.5387Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M13.7846 69.5082C12.7698 69.5078 11.9469 70.3307 11.9465 71.3455C11.9461 72.3602 12.7689 73.1831 13.7837 73.1835C14.7985 73.1839 15.6214 72.3611 15.6218 71.3463C15.6205 70.3324 14.7989 69.5099 13.7846 69.5082Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M3.2489 64.9678C3.22946 65.2091 3.25016 65.4517 3.3106 65.6862C3.36512 65.9233 3.4412 66.155 3.53756 66.3785C5.06162 65.7657 6.66132 64.942 6.36251 61.5503C6.25051 60.7591 6.09836 59.9742 5.90606 59.1987C4.5481 53.6477 3.98599 47.9314 4.23619 42.2219C4.65884 33.7682 8.83202 25.981 10.8383 22.2385C11.2559 21.4591 11.5864 20.8442 11.7466 20.4723C12.3421 19.0987 9.84763 16.9685 8.83117 16.2564C8.36246 16.5467 8.03238 17.0163 7.91784 17.5556C7.83881 17.9009 7.85571 18.261 7.96729 18.597C4.95722 23.1053 2.77934 28.1162 1.53719 33.3929C1.53719 33.3929 1.47971 33.6063 1.42646 33.8878C0.62428 37.4608 0.221501 41.1121 0.225728 44.7743C0.191916 44.8491 0.239252 46.4306 0.257849 46.8625C0.138663 53.0298 1.15258 59.1665 3.2489 64.9678Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M50.3225 23.9734C49.3978 23.973 48.6476 24.7228 48.6476 25.6475C48.6472 26.5723 49.3969 27.3225 50.3217 27.3225C51.2464 27.3225 51.9966 26.5731 51.9966 25.6484C51.9953 24.7241 51.2468 23.9747 50.3225 23.9734Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M40.1157 51.6481C41.0115 51.6481 41.7378 50.9219 41.7378 50.026C41.7378 49.1302 41.0115 48.4039 40.1157 48.4039C39.2198 48.4039 38.4936 49.1302 38.4936 50.026C38.4936 50.9219 39.2198 51.6481 40.1157 51.6481Z"
+          fill="#0B0C0C"
+        />
+        <path
+          d="M54.2176 19.9744C55.0257 17.0666 54.1753 15.1017 53.3182 13.9602C52.8119 13.2784 52.1657 12.7129 51.4227 12.3021C48.1814 10.0968 44.2212 11.5769 44.1029 11.6314C42.1866 12.9251 39.8076 13.3351 37.5688 12.7573C31.5309 11.2164 26.7787 3.47607 25.8865 1.94019L25.3865 1.33707C25.242 1.17562 25.048 1.06743 24.8346 1.02854L23.6693 0.267785C23.3803 0.0708328 22.9978 0.0805533 22.7192 0.291875L22.6009 0.379785C22.5903 0.379785 22.5806 0.374714 22.5684 0.373023C21.8173 0.303287 21.0667 0.513764 20.4615 0.964725C19.7007 1.46598 18.9611 2.26182 19.0444 3.04117L17.5165 4.18569C16.8982 4.23641 16.3056 4.45365 15.801 4.81458C15.1945 5.20891 14.5546 5.83907 14.4122 6.50812L13.4198 7.25324L13.2571 7.39567L13.2385 7.48358C13.1861 7.566 13.1497 7.65729 13.132 7.75323C13.1075 7.81155 13.0969 7.87495 13.1011 7.93792L13.1134 8.18517L13.1844 8.27308C13.2981 8.50004 14.0449 9.99239 14.5364 10.9205C12.9625 12.3533 11.5069 13.9107 10.1841 15.578C11.4545 16.5814 13.9143 18.8768 12.99 21.0099C12.8066 21.4325 12.4828 22.039 12.0327 22.8779C10.0733 26.533 5.99523 34.1393 5.58907 42.2896C5.3469 47.8777 5.89887 53.4723 7.22893 58.9058C7.43645 59.738 7.5979 60.5811 7.71201 61.4315C8.07633 65.5599 5.91747 66.8456 4.29706 67.5294C5.52864 70.3142 7.1013 72.9354 8.97783 75.3335C10.3869 76.9425 12.2736 79.942 11.8894 81.7285C11.4465 81.5671 10.9689 81.5236 10.504 81.6017C9.61264 81.8029 8.86752 82.4103 8.49095 83.2429C7.94785 84.4284 8.17692 85.8244 9.07039 86.7741C9.86073 87.5416 11.0395 87.7174 12.4722 87.2432C12.4845 87.2724 12.4989 87.3002 12.5145 87.3277C12.6688 87.6079 12.9486 87.7969 13.2664 87.8349C13.4988 87.8607 13.7321 87.8738 13.9663 87.8738C16.544 87.8738 20.6373 86.3396 25.578 79.0824C27.302 79.3939 28.7977 79.6259 30.0056 79.77L31.9582 80.1124C32.7765 80.3799 34.5549 80.278 35.4674 77.9628L55.163 27.0177C56.2564 24.6881 55.8883 21.9346 54.2214 19.974L54.2176 19.9744ZM25.5045 5.40586C25.5467 5.4794 29.8213 12.8135 38.0118 14.4906C38.5498 14.603 39.0975 14.6609 39.647 14.6639C41.3574 14.6639 42.6545 14.0489 43.9097 13.4543C45.0889 12.8955 46.1979 12.3706 47.5403 12.3706C47.9181 12.3715 48.2947 12.4129 48.6636 12.4949C52.7028 13.3824 52.9818 17.6012 52.9843 17.6435C53.0038 17.9698 52.7553 18.25 52.429 18.2694C52.4285 18.2694 52.4294 18.2694 52.429 18.2694H52.3918C52.079 18.2686 51.8204 18.0256 51.8001 17.7132C51.7912 17.5755 51.5503 14.3418 48.4088 13.6521C48.1243 13.5887 47.8336 13.5565 47.5419 13.5561C46.4608 13.5561 45.5137 14.005 44.4186 14.5244C43.1088 15.1448 41.6266 15.8477 39.641 15.8477C39.0134 15.8452 38.3879 15.7792 37.7734 15.6507C29.0331 13.8604 24.5155 6.0673 24.4711 5.98911C24.3092 5.70425 24.4085 5.34247 24.6934 5.18017C24.9783 5.01788 25.3401 5.11762 25.5023 5.40248L25.5045 5.40586ZM40.4965 23.8103V20.6912L41.2839 20.4376C41.548 20.3429 41.7847 20.2521 41.9812 20.1713L41.9964 21.2905L40.4965 23.8103ZM40.5163 23.8885V29.1124L38.9699 31.9597V26.5021L40.5163 23.8885ZM34.9882 33.0738L32.623 37.0517V45.8368L31.077 48.6841V39.6518L32.6163 37.0475V21.3049L32.929 21.3569C33.3094 21.4105 33.683 21.4414 34.0524 21.4659L34.9932 21.5001H35.0101V33.0358L36.5566 30.4209V21.4456L37.1766 21.3974C37.5194 21.3573 37.8423 21.3044 38.1614 21.2491L38.9504 21.091V26.4104L36.5557 30.4357V37.6992L35.0093 40.5466L35.0101 33.0358L34.9882 33.0738ZM24.7353 50.3147V18.6452C25.0049 18.8354 25.2851 19.0099 25.575 19.1684L26.5163 19.6333C26.7103 19.7178 26.9152 19.8058 27.1295 19.8907V46.2523L28.6756 43.6369V20.43L28.6874 20.4342C29.0411 20.5416 29.4059 20.6455 29.7947 20.7453L30.7351 20.9604C30.8437 20.9837 30.9574 21.0052 31.0694 21.0276V39.662L28.6756 43.6877V53.9938L27.1295 56.8416V46.289L24.7353 50.3147ZM20.7937 56.9405V19.6063L21.4183 18.9808L23.0409 17.3481C23.0831 17.3713 23.1389 17.4034 23.1905 17.4326V52.9148L21.3406 56.0259L20.7937 56.9405ZM20.8664 56.9557V69.0666L19.3203 71.9139V59.5706L20.8664 56.9557ZM23.1875 53.0264L24.734 50.4111V62.5219L23.1875 65.3697V53.0264ZM16.2046 9.04313L24.0916 1.96343C24.1998 1.86538 24.3409 1.81128 24.4872 1.81128C24.8151 1.81212 25.0801 2.07839 25.0797 2.40636C25.0797 2.41651 25.0793 2.42665 25.0789 2.43679C25.0708 2.59402 25.0002 2.74152 24.8823 2.84633L17.3348 9.62216C18.2164 11.2848 19.4949 12.7041 21.0574 13.7531C21.3355 13.9259 21.4205 14.2919 21.2476 14.57C21.0747 14.8481 20.7087 14.9331 20.4306 14.7602C18.5478 13.5058 17.0376 11.7675 16.0588 9.72782C15.9531 9.49198 16.0123 9.21515 16.2046 9.04313ZM16.3935 13.8976C18.0486 13.8976 19.3901 15.2391 19.3901 16.8942C19.3901 18.5492 18.0486 19.8907 16.3935 19.8907C14.7385 19.8907 13.397 18.5492 13.397 16.8942C13.3987 15.2404 14.7393 13.8997 16.3935 13.8976ZM19.2485 21.241V59.541L19.2307 59.571L15.8082 65.3253C11.0222 56.3471 10.8442 47.0506 10.8425 46.9374C10.31 34.8726 15.0432 26.3542 19.2485 21.241ZM10.5949 71.3455C10.5953 69.5839 12.0234 68.1558 13.785 68.1562C15.5466 68.1566 16.9747 69.5847 16.9742 71.3463C16.9738 73.1079 15.5461 74.5356 13.7846 74.5356C12.0238 74.5339 10.5966 73.1066 10.5949 71.3455ZM20.1551 75.8901L14.9063 86.2724C14.8057 86.4723 14.6003 86.5982 14.3763 86.5978C14.2837 86.5978 14.192 86.5758 14.1091 86.5344C13.8171 86.3878 13.6992 86.0323 13.8458 85.7403C13.8463 85.739 13.8471 85.7382 13.8475 85.7369L19.0963 75.3551C19.1969 75.1552 19.4019 75.0288 19.6259 75.0292C19.9526 75.0279 20.2185 75.2917 20.2197 75.6184C20.2201 75.713 20.1978 75.806 20.1551 75.8901ZM24.8278 76.3643C24.7437 76.5727 24.5611 76.7257 24.3409 76.7726C24.1207 76.8191 23.8921 76.7532 23.7302 76.5968C23.7209 76.5879 22.7619 75.6674 19.912 73.3306C19.623 73.0939 19.5807 72.6679 19.8178 72.3788C20.0545 72.0897 20.4805 72.0475 20.7696 72.2846C22.3418 73.5724 23.3435 74.4333 23.9386 74.9586C26.6346 68.2897 43.3755 26.8605 44.3987 23.9278C45.2761 21.4126 46.5073 20.381 47.386 19.9587C48.1019 19.5936 48.9358 19.5353 49.6957 19.7969C50.0478 19.9211 50.2325 20.3074 50.1082 20.6595C49.984 21.0116 49.5977 21.1963 49.2456 21.072C48.8242 20.9338 48.3648 20.9718 47.9722 21.1777C47.0563 21.618 46.2398 22.7524 45.6755 24.3728C44.5564 27.5777 25.6325 74.3737 24.8278 76.3626V76.3643ZM47.5846 29.5401C48.0707 29.5401 48.465 29.934 48.465 30.4204C48.465 30.9065 48.0711 31.3008 47.5846 31.3008C47.0986 31.3008 46.7043 30.9069 46.7043 30.4204C46.7038 29.9344 47.0986 29.5401 47.5846 29.5401C47.5846 29.5397 47.5842 29.5401 47.5846 29.5401ZM45.2499 36.0327C45.2525 36.3599 44.99 36.6274 44.6629 36.6316H44.6582C44.3337 36.6299 44.0708 36.3687 44.0665 36.0441L44.0044 30.7201C44.001 30.3926 44.2639 30.1246 44.5915 30.1212C44.919 30.1178 45.187 30.3807 45.1904 30.7083L45.2499 36.0327ZM43.0133 36.848C43.0129 37.3341 42.619 37.7275 42.1329 37.7275C41.6469 37.7275 41.2534 37.3332 41.2534 36.8472C41.2539 36.3615 41.6473 35.9676 42.1334 35.9676C42.6194 35.9668 43.0137 36.3599 43.015 36.8459C43.015 36.8467 43.015 36.8472 43.015 36.848H43.0133ZM37.6411 54.5103C38.1271 54.5099 38.5215 54.9038 38.5219 55.3902C38.5223 55.8767 38.1284 56.2706 37.6419 56.271C37.1559 56.2714 36.7616 55.8775 36.7612 55.3911C36.7612 55.3907 36.7612 55.3915 36.7612 55.3911C36.7607 54.905 37.1551 54.5107 37.6411 54.5103ZM28.6464 71.0737C28.4427 71.3302 28.0691 71.3733 27.8125 71.1692C27.556 70.9655 27.5129 70.5919 27.717 70.3353L32.7536 63.9936C32.8661 63.8516 33.0372 63.7687 33.2186 63.7691C33.5453 63.767 33.8119 64.0299 33.8141 64.3566C33.8149 64.4931 33.7688 64.6254 33.6835 64.7315L28.6464 71.0737ZM32.1958 71.5293C32.1958 72.0153 31.8019 72.4092 31.3158 72.4092C30.8298 72.4092 30.4359 72.0153 30.4359 71.5293C30.4359 71.0433 30.8298 70.6494 31.3158 70.6494C31.8019 70.6494 32.1958 71.0428 32.1958 71.5293ZM30.9764 61.9704C30.976 61.4843 31.3699 61.09 31.8564 61.0896C32.3424 61.0892 32.7367 61.4831 32.7372 61.9695C32.7376 62.4556 32.3437 62.8499 31.8572 62.8503C31.8568 62.8503 31.8576 62.8503 31.8572 62.8503C31.3716 62.8499 30.9769 62.456 30.9764 61.9704ZM29.9025 76.7866C29.4165 76.787 29.0221 76.3931 29.0217 75.9066C29.0213 75.4202 29.4152 75.0263 29.9016 75.0258C30.3877 75.0254 30.782 75.4193 30.7824 75.9058C30.7824 76.3918 30.3881 76.7861 29.9025 76.7866ZM33.7249 77.9755C33.6809 78.2472 33.4256 78.4319 33.1539 78.388C32.9121 78.3491 32.7346 78.1407 32.7351 77.896C32.7359 77.8415 32.7456 77.7874 32.7638 77.7358C33.0926 76.0448 31.9887 74.4079 30.2977 74.0791C28.6227 73.7533 26.9972 74.8339 26.649 76.5042C26.6439 76.5186 26.638 76.533 26.6308 76.5465C26.5484 76.8085 26.2694 76.9544 26.0074 76.8719C25.8003 76.8068 25.6591 76.615 25.6591 76.3981C25.6591 76.3893 25.6592 76.3808 25.6617 76.3719C25.6587 76.3398 25.6604 76.3069 25.6668 76.2752C26.1583 74.0542 28.3569 72.6523 30.5779 73.1434C32.768 73.6282 34.1678 75.7764 33.7261 77.9759L33.7249 77.9755ZM34.4679 55.0893C34.7921 55.0948 35.0549 55.3526 35.0668 55.6768L35.1289 61.0723C35.1331 61.399 34.8715 61.6673 34.5448 61.6716C34.544 61.6716 34.5431 61.6716 34.5423 61.6716H34.5359C34.2118 61.6694 33.9489 61.4087 33.9442 61.0845L33.8821 55.6886C33.8796 55.3619 34.1408 55.094 34.4679 55.0893ZM33.8969 63.1601C33.8973 62.6741 34.2912 62.2806 34.7773 62.2806C35.2633 62.281 35.6568 62.6749 35.6568 63.161C35.6564 63.6466 35.2629 64.0405 34.7768 64.0405C34.2912 64.0401 33.8969 63.6462 33.8969 63.1601ZM35.5228 74.8644C35.5144 75.186 35.2527 75.4426 34.9311 75.4455H34.9243C34.5955 75.4404 34.3326 75.1695 34.3377 74.8407L34.5321 65.229C34.5431 64.9061 34.8073 64.6495 35.1306 64.6478C35.4594 64.652 35.7227 64.9217 35.7185 65.2505C35.7185 65.2514 35.7185 65.2522 35.7185 65.2526L35.5228 74.8644ZM37.6567 65.1584C37.1707 65.1584 36.7764 64.7645 36.7764 64.278C36.7764 63.792 37.1703 63.3976 37.6567 63.3976C38.1432 63.3976 38.5371 63.7915 38.5371 64.2776C38.5367 64.7632 38.1428 65.1567 37.6572 65.1567L37.6567 65.1584ZM42.4808 55.878L36.4184 62.0989C36.1897 62.3334 35.814 62.3385 35.5794 62.1094C35.3449 61.8808 35.3398 61.505 35.5689 61.2705L41.6325 55.0492C41.8612 54.8146 42.2365 54.8095 42.4715 55.0382C42.7061 55.2668 42.7111 55.6421 42.4825 55.8771L42.4808 55.878ZM40.1148 53.0019C38.472 53.0019 37.1403 51.6701 37.1403 50.0273C37.1403 48.3845 38.472 47.0527 40.1148 47.0527C41.7576 47.0527 43.0894 48.3845 43.0894 50.0273C43.0869 51.6689 41.7568 52.9989 40.1157 53.001L40.1148 53.0019ZM42.3012 45.5443C41.8151 45.5447 41.4208 45.1508 41.4204 44.6648C41.42 44.1788 41.8139 43.7844 42.3003 43.784C42.7864 43.7836 43.1807 44.1775 43.1811 44.664C43.1807 45.1492 42.7876 45.5431 42.302 45.5435L42.3012 45.5443ZM43.8062 39.6928L39.3203 45.341C39.1165 45.5976 38.7434 45.6407 38.4864 45.437C38.2298 45.2333 38.1867 44.8601 38.3904 44.6031L42.8764 38.9545C42.9888 38.8125 43.16 38.73 43.3413 38.7305C43.668 38.7284 43.9347 38.9917 43.9364 39.3184C43.9372 39.4545 43.8911 39.5867 43.8062 39.6928ZM43.9541 37.9452C43.9537 37.4592 44.3476 37.0648 44.8341 37.0644C45.3205 37.064 45.7144 37.4579 45.7149 37.9443C45.7153 38.4308 45.3214 38.8247 44.8349 38.8251C44.8345 38.8251 44.8353 38.8251 44.8349 38.8251C44.3489 38.8247 43.9545 38.4304 43.955 37.9443L43.9541 37.9452ZM45.6865 49.1157C45.6781 49.4373 45.4165 49.6938 45.0948 49.6968H45.0877C44.7588 49.6913 44.4964 49.4208 44.501 49.092L44.6773 40.3788C44.6785 40.3335 44.6853 40.2887 44.6967 40.2448V40.1903C44.7018 40.0233 44.8294 39.8856 44.9955 39.8674C45.0809 39.8226 45.176 39.7985 45.2723 39.7981C45.6016 39.8002 45.8662 40.069 45.864 40.3978C45.864 40.3995 45.864 40.4007 45.864 40.4024L45.6865 49.1157ZM47.479 39.883C46.9929 39.8834 46.5986 39.4895 46.5982 39.0031C46.5978 38.5166 46.9917 38.1227 47.4781 38.1223C47.9642 38.1219 48.3585 38.5158 48.3585 39.0022C48.3576 39.4874 47.9646 39.8809 47.4794 39.8817L47.479 39.883ZM52.6263 31.0286L46.5627 37.2495C46.3353 37.4841 45.9604 37.4896 45.7263 37.2618C45.7258 37.2614 45.725 37.2605 45.7246 37.2601C45.5657 37.1088 45.5065 36.881 45.5712 36.6713C45.5606 36.5737 45.5936 36.4761 45.6616 36.4051L51.7253 30.1842C51.7806 30.1288 51.8538 30.095 51.9319 30.0887C52.2219 29.9378 52.579 30.0506 52.7299 30.3406C52.8478 30.568 52.806 30.8456 52.6263 31.0286ZM50.3217 28.6754C48.6497 28.6758 47.2943 27.3204 47.2943 25.6488C47.2938 23.9768 48.6493 22.6214 50.3208 22.6214C51.9928 22.621 53.3482 23.9764 53.3482 25.648C53.3482 25.6475 53.3482 25.6484 53.3482 25.648C53.3457 27.3182 51.9924 28.6724 50.3225 28.6745L50.3217 28.6754Z"
+          fill="#0B0C0C"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_212_1062">
+          <rect width="56" height="88" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export default HarpBlack;
